# contracts/openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: OrlyFin API
  version: 1.3.0
  description: API REST versionada para OrlyFin.
servers:
  - url: /api/v1
paths:
...
  /reconciliations:
    post:
      summary: Create reconciliation for account period
      tags: [reconciliation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReconciliationCreateRequest'
      responses:
        '201':
          description: Reconciliation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationResponse'
    get:
      summary: List reconciliations
      tags: [reconciliation]
      responses:
        '200':
          description: Reconciliation list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationListResponse'
  /reconciliations/{reconciliation_id}/resolve:
    post:
      summary: Resolve reconciliation
      tags: [reconciliation]
      parameters:
        - in: path
          name: reconciliation_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReconciliationResolveRequest'
      responses:
        '200':
          description: Reconciliation resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationResponse'
...
  /attachments/upload:
    post:
      summary: Upload attachment
      tags: [attachments]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AttachmentUploadRequest'
      responses:
        '200':
          description: Attachment uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentResponse'
...
  /attachments/{attachment_id}:
    delete:
      summary: Delete attachment
      tags: [attachments]
      parameters:
        - in: path
          name: attachment_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Attachment deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentResponse'
  /settings:
    post:
      summary: Upsert setting
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingUpsertRequest'
      responses:
        '200':
          description: Setting updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingResponse'
    get:
      summary: List settings
      tags: [settings]
      responses:
        '200':
          description: Settings list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingListResponse'
  /settings/{key}:
    get:
      summary: Get setting by key
      tags: [settings]
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Setting detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingResponse'
components:
  schemas:
...
    ReconciliationCreateRequest:
      type: object
      required: [account_id, period_start, period_end, statement_balance]
      properties:
        account_id:
          type: string
        period_start:
          type: string
          format: date
        period_end:
          type: string
          format: date
        statement_balance:
          type: number
        notes:
          type: string
    ReconciliationResolveRequest:
      type: object
      properties:
        notes:
          type: string
    ReconciliationResponse:
      type: object
      properties:
        id:
          type: string
        account_id:
          type: string
        period_start:
          type: string
          format: date
        period_end:
          type: string
          format: date
        book_balance:
          type: number
        statement_balance:
          type: number
        difference_amount:
          type: number
        status:
          type: string
        notes:
          type: string
        resolved_on:
          type: string
          nullable: true
    ReconciliationListResponse:
      type: object
      properties:
        reconciliations:
          type: array
          items:
            $ref: '#/components/schemas/ReconciliationResponse'
...
    AttachmentUploadRequest:
      type: object
      required: [module, entity_id, file]
      properties:
        module:
          type: string
        entity_id:
          type: string
        file:
          type: string
          format: binary
    AttachmentResponse:
      type: object
      properties:
        id:
          type: string
        module:
          type: string
        entity_id:
          type: string
        file_name:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
        status:
          type: string
...
    SettingUpsertRequest:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
            - type: object
            - type: array
        value_type:
          type: string
        is_sensitive:
          type: boolean
    SettingResponse:
      type: object
      properties:
        key:
          type: string
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
            - type: object
            - type: array
            - type: 'null'
        value_type:
          type: string
        is_sensitive:
          type: boolean
    SettingListResponse:
      type: object
      properties:
        settings:
          type: array
          items:
            $ref: '#/components/schemas/SettingResponse'
...